Make the email field an `email` field.
This asks the browser to do whatever validation it thinks is suitable
for email addresses. This is much better than Joomla's kinda-sorta
working filter. It doesn't prevent users from entering something totally
bogus. At least it hints phones and tables to use a keyboard interface
which is more appropriate for entering emails...

Closes gh-34