Shared Sessions can lead to comments not being displayed
If the manager applies filters in the backend Comments page when
using the Shared Sessions feature in Joomla then these filters are
applied to the frontend as well.

The issue is addressed by using a temporary model instance in the
frontend Html class, setting the asset_id model state manually. The
temporary model instance does not honor the (shared) session state,
circumventing the problem.