Checkbox to accept ToS / give consent
Close gh-46

Akeeba Engage can be configured so that Guest users can post comments.

Guest users have not necessarily accepted the site's privacy policy but we do collect their name and email address which is personally identifiable information. The GDPR says that the user must provide explicit and positive action to indicate that they consent to their information being collected. Depending on who you ask, submitting the form is such an action or not.

If you or your country believes that someone filling in a form and submitting it on their own free will is not performing an explicit and positive action you need a checkbox to provide explicit consent, one which is always unchecked unless the user deliberately goes in and checks it.

This feature adds this kind of checkbox. Moreover, it lets you enter full HTML and plugin tags for the displayed message so you can do all sorts of formatting, including implementing popups with third party plugins.

Note: the checkbox is **only ever** shown to Guests. You MUST obtain the consent of logged in users when they register a user account or, if it's an account created before GDPR came into effect, when they first login. You can use Joomla's `com_privacy` for that. Since anyone who's logged in already has (or should have...) provided consent we do NOT show this checkbox to them.