# [MEDIUM] Joomla no longer supports multi–state toggles
This was causing the XML form display issues in #159 and
https://github.com/akeeba/engage/discussions/141#discussioncomment-1934343

This is a really STUPID Joomla issue. In previous versions you could
have a multistate toggle e.g. Yes / No / Use Global. If you try to do
that as of Joomla 4.0.5 you get a broken binary state (Yes / No) toggle.
The only solution is to use a stupid dropdown like we're back in the
olden Joomla 1.5 days. I fail to see how this is an improvement in
Joomla BUT they are using this shitty display everywhere now, so...