# [HIGH] The Email plugin never fired
I believe that's the problem with #141.

I am not sure why he's getting a PHP error, though. The actual problem
is that the plugin never fired because it was not implementing
SubscriberInterface and the event name was wrong (merge failure with the
branch I was developing the plugin).

It turns out there were a couple more issues, mostly affecting sending
emails to registered users.

The most annoying issue was not sending emails to users with Receive
System Emails set to No. While makes sense respecting Joomla's flag, the
way it is implemented in Joomla makes no sense. New Joomla users —
including Super Users — have that set to No by default, making it a
useless flag for the purpose Engage sends emails :(