~ Prevent comments display when “Show unauthorised links” option is enabled in Joomla's content component or menu item
This hides the comment count and the
comments display / new comment
interface based on the article's access
level.

This is necessary because Joomla will
_idiotically_ call the plugin events
when rendering an “unauthorised” link,
i.e. when displaying the introtext of
an article the current user has no
access to. Joomla will _idiotically_
only remove the `fulltext` of the
article in this case which is
completely useless for content plugins.
Joomla's stupidity is so bad that
even its own custom fields will
blithely render in this case, leaking
privileged information to the public.
YOLO!

Close gh-238